+++ java/client/src/main/java/glide/api/GlideClusterClient.java
@@ -110,17 +110,31 @@ public class GlideClusterClient extends BaseClient
     public CompletableFuture<ClusterValue<Object>> customCommand(
             @NonNull String[] args, @NonNull Route route) {
         return commandManager.submitNewCommand(
-                CustomCommand, args, route, response -> handleCustomCommandResponse(route, response));
+                CustomCommand,
+                args,
+                route,
+                response -> handleCustomCommandResponse(route, response, false));
     }
 
-    protected ClusterValue<Object> handleCustomCommandResponse(Route route, Response response) {
+    // @Override
+    public CompletableFuture<ClusterValue<Object>> customCommand(
+            @NonNull GlideString[] args, @NonNull Route route) {
+        return commandManager.submitNewCommand(
+                CustomCommand, args, route, response -> handleCustomCommandResponse(route, response, true));
+    }
+
+    protected ClusterValue<Object> handleCustomCommandResponse(
+            Route route, Response response, boolean bin) {
         if (route instanceof SingleNodeRoute) {
-            return ClusterValue.ofSingleValue(handleObjectOrNullResponse(response));
+            return ClusterValue.ofSingleValue(
+                    bin ? handleBinaryObjectOrNullResponse(response) : handleObjectOrNullResponse(response));
         }
         if (response.hasConstantResponse()) {
             return ClusterValue.ofSingleValue(handleStringResponse(response));
         }
-        return ClusterValue.ofMultiValue(handleMapResponse(response));
+        return bin
+                ? ClusterValue.ofMultiValueBinary(handleBinaryStringMapResponse(response))
+                : ClusterValue.ofMultiValue(handleMapResponse(response));
     }
 
     @Override